WIP: Create useConsolidatedRef hook
