+++ packages/main/src/hooks/useConsolidatedRef.ts
@@ -1,0 +1,21 @@
+import { RefObject, useEffect, useMemo, useRef } from 'react';
+
+export const useConsolidatedRef = <T>(ref): RefObject<T> => {
+  const localPopoverRef: RefObject<T> = useRef(null);
+
+  const popRef = useMemo(() => {
+    if (!ref || typeof ref === 'function') {
+      return localPopoverRef;
+    }
+    return ref;
+  }, [ref]);
+
+  useEffect(() => {
+    if (typeof ref === 'function') {
+      // @ts-ignore
+      givenRef(popRef.current);
+    }
+  }, [popRef.current]);
+
+  return popRef;
+};
